fix: nearcore pytests do not play nice with common cluster, simple fix until we have a proper handling of this in a follow-up
