+++ pytest/tests/test_lost_assets.py
@@ -28,7 +28,7 @@ PRESIGNATURES_TO_BUFFER = 8
 @pytest.fixture(scope="module")
 def lost_assets_cluster():
     """
-    Spins up a cluster with three nodes, initializes the contract and adds domains. Retuns the cluster in a running state.
+    Spins up a cluster with three nodes, initializes the contract and adds domains. Returns the cluster in a running state.
     """
     cluster, mpc_nodes = shared.start_cluster_with_mpc(
         2,
